Speed up the admin dashboard's load time
This allows the page to load quickly and data to periodically refresh
in the background.

The following sections have been modified to load via Ajax and refresh
on a sceduled basis.  The loading is delayed a random amount up to 2
seconds and then each section updates every 15-20 minutes.  The reload
interval for each section is set in the initialize method of dc.ui.Admin

  * Statistics
  * Charts
  * Most recent docs
  * Break down by the numbers

The remaining sections of the page are static and will not refresh.