Resources access can be INVISIBLE and DELETED
Annotations that are INVISIBLE or DELETED may only be viewed or edited by Accounts that have the ADMINISTRATOR or CONTRIBUTOR roles.

A bug was fixed in the annotations & api controllers that was allowing access to annotations regardless of what the access level was set to.

The viewer was updated to include support for Annotations with the additional access levels as well.