Detect and fix uploaded PDF's with incorrect extensions
If a file is uploaded that has an extension that is a PDF but has
an incorrect extension we would end up converting from PDF to PDF.
This uses the file command which is available on all platforms to
check the mime-type of the file and only convert if it's not a
PDF.  If it is a PDF with incorrect extension it simply renames
the file to have a '.pdf' extension.

There is another case; the document could be a non pdf
with an incorrect extension.  For instance .docx but named
.xlsx (stranger things have happened)

In that case, LibreOffice is smart enough to save us.  Somehow it can
silently handle any supported file type being thrown at it and then
convert it to PDF.  I've tested this behavior with both the existing
JODConvertor and with my work replacing it with
unoconv (https://github.com/nathanstitt/docsplit).